creating a notification service
